+++ NEWS.md
@@ -4,7 +4,7 @@
 * Added "format" parameter to `datastep()` function.
 * Added `lib_export()` function to export a library to a different 
 directory and file format.
-* Fixed bug on `dictionary()` that was causing it to crash when there 
+* Fixed bug on `dictionary()` that was causing a warning when there 
 were no rows on the input data frame.
 
 # libr 1.2.2